Move NCore to small and move linear.
