+++ docs/source/polymer.rst
@@ -6,4 +6,3 @@ Polymers and Small Molecules
     :maxdepth: 1
 
     Linear <_autosummary/stk.polymer.Linear>
-    NCore <_autosummary/stk.polymer.NCore>
