+++ src/stk/_internal/topology_graphs/small/vertices.py
@@ -1,12 +1,16 @@
+"""
+Small Molecule Vertices
+=======================
+
+"""
+
 import logging
 
 import numpy as np
 
-from stk._internal.building_block import BuildingBlock
 from stk._internal.topology_graphs.vertex import Vertex
 from stk._internal.utilities.utilities import get_acute_vector
 
-from ..edge import Edge
 from ..utilities import _EdgeSorter, _FunctionalGroupSorter
 
 logger = logging.getLogger(__name__)
@@ -18,11 +22,7 @@ class CoreVertex(Vertex):
 
     """
 
-    def place_building_block(
-        self,
-        building_block: BuildingBlock,
-        edges: tuple[Edge, ...],
-    ) -> np.ndarray:
+    def place_building_block(self, building_block, edges):
         # Sets building block in XY plane, then aligns 0th functional
         # group with the 0th edge.
         assert building_block.get_num_functional_groups() > 2, (
@@ -46,7 +46,7 @@ class CoreVertex(Vertex):
         )
         building_block = building_block.with_rotation_between_vectors(
             start=normal,
-            target=np.array([0.0, 0.0, 1.0]),
+            target=[0, 0, 1],
             origin=self._position,
         )
         (fg,) = building_block.get_functional_groups(0)
@@ -59,11 +59,10 @@ class CoreVertex(Vertex):
             origin=self._position,
         ).get_position_matrix()
 
-    def map_functional_groups_to_edges(
-        self,
-        building_block: BuildingBlock,
-        edges: tuple[Edge, ...],
-    ) -> dict[int, int]:
+    # def map_functional_groups_to_edges(self, building_block, edges):
+    #     return {fg_id: edge.get_id() for fg_id, edge in enumerate(edges)}
+
+    def map_functional_groups_to_edges(self, building_block, edges):
         # The idea is to order the functional groups in building_block
         # by their angle with the vector running from the placer
         # centroid to fg0, going in the clockwise direction.
@@ -94,11 +93,7 @@ class SubstituentVertex(Vertex):
 
     """
 
-    def place_building_block(
-        self,
-        building_block: BuildingBlock,
-        edges: tuple[Edge, ...],
-    ) -> np.ndarray:
+    def place_building_block(self, building_block, edges):
         assert building_block.get_num_functional_groups() == 1, (
             f"{building_block} needs to have 1 functional "
             "groups but has "
@@ -124,9 +119,5 @@ class SubstituentVertex(Vertex):
             origin=self._position,
         ).get_position_matrix()
 
-    def map_functional_groups_to_edges(
-        self,
-        building_block: BuildingBlock,
-        edges: tuple[Edge, ...],
-    ) -> dict[int, int]:
+    def map_functional_groups_to_edges(self, building_block, edges):
         return {0: edges[0].get_id()}
