+++ src/stk/polymer.py
@@ -1,10 +1,7 @@
-from stk._internal.topology_graphs.polymer.linear.linear import Linear
-from stk._internal.topology_graphs.polymer.ncore.ncore import NCore
+from stk._internal.topology_graphs.polymer.linear import Linear
 from stk._internal.topology_graphs.polymer.vertices import (
-    CoreVertex,
     HeadVertex,
     LinearVertex,
-    SubstituentVertex,
     TailVertex,
     TerminalVertex,
     UnaligningVertex,
@@ -12,12 +9,9 @@ from stk._internal.topology_graphs.polymer.vertices import (
 
 __all__ = [
     "Linear",
-    "NCore",
     "TerminalVertex",
     "HeadVertex",
     "TailVertex",
     "LinearVertex",
     "UnaligningVertex",
-    "CoreVertex",
-    "SubstituentVertex",
 ]
