+++ tests/molecular/topology_graphs/vertex/cases/polymer/conftest.py
@@ -11,8 +11,6 @@ from ...case_data import CaseData
         lazy_fixture("center"),
         lazy_fixture("head"),
         lazy_fixture("tail"),
-        lazy_fixture("substituent"),
-        lazy_fixture("core"),
     ),
 )
 def case_data(request: pytest.FixtureRequest) -> CaseData:
@@ -72,32 +70,6 @@ def unaligning(
             id=id,
             position=position,
             flip=flip,
-        ),
-        id=id,
-        position=position,
-        cell=np.array([0, 0, 0]),
-    )
-
-
-@pytest.fixture
-def core(id, position, flip):
-    return CaseData(
-        vertex=stk.polymer.CoreVertex(
-            id=id,
-            position=position,
-        ),
-        id=id,
-        position=position,
-        cell=np.array([0, 0, 0]),
-    )
-
-
-@pytest.fixture
-def substituent(id, position, flip):
-    return CaseData(
-        vertex=stk.polymer.SubstituentVertex(
-            id=id,
-            position=position,
         ),
         id=id,
         position=position,
